requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
redis-om==0.2.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.24.3
typing-extensions==4.8.0